backend/requirements.txt
pydantic-ai==0.0.33
python-dotenv==1.0.1